Restructure of libbri.  Everything is now included in bri.h or bri++.H, with all objects in libbri.a.  Moved an interval list like class from the current three search programs (searchGENOME, seatac and snapper2) to merCovering.H, in bri++.H.  Import of Mersenne Twister pseudo-RNG.  Lots of files had changes to just the list of included files.

+++ kmer/libsim4/sim4core/pluri_align.C
@@ -174,7 +174,7 @@ Sim4::pluri_align(int *dist_ptr,
 
     diff = align_get_dist(nextExon->frGEN-1, nextExon->frEST-1,
                           nextExon->toGEN, nextExon->toEST,
-                          max(1000, .2*(nextExon->toEST - nextExon->frEST + 1)));
+                          max(1000, (int)(.2*(nextExon->toEST - nextExon->frEST + 1))));
 
 
     //  Return if the alignment fails.
